runtime : add Messages.h header
